requirement_20250819.txt
asyncpg==0.30.0
greenlet==3.2.4
numpy==2.3.2
pandas==2.3.1
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
tzdata==2025.2
